move next date to utils
