+++ internal/utils/next_date.go
@@ -1,4 +1,4 @@
-package models
+package utils
 
 import (
 	"errors"
